pyproject.toml
[project]
name = "build-optimiser"
version = "0.2.0"
description = "Analyse C++ codebases to optimise build times via CMake File API, dependency graph analysis, and build metric collection"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    # Core data handling
    "pandas>=2.0",
    "pyarrow>=14.0",
    # Graph analysis
    "networkx>=3.0",
    # Community detection
    "python-louvain>=0.16",
    "leidenalg>=0.10",
    "igraph>=0.11",
    # Clustering and dimensionality reduction
    "scikit-learn>=1.3",
    "umap-learn>=0.5",
    # Statistical analysis and hierarchical clustering
    "scipy>=1.11",
    # Graph partitioning
    "pymetis>=2023.1",
    # Visualisation
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "pyvis>=0.3",
    "adjustText>=1.1",
    # Configuration
    "pyyaml>=6.0",
    # Schema validation
    "pandera>=0.20",
    # Notebooks
    "jupyterlab>=4.0",
    # Code metrics
    "pygments>=2.0",
    "missingno>=0.5.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring CMake, GCC, and Ninja (deselect with '-m \"not integration\"')",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=7.0",
    "ruff>=0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/buildanalysis"]
